WIIS-1772 : on supprime prise & dépose pour la modif + gestion des PJ + refacto pour persistMouvementTraca

+++ src/Service/AttachmentService.php
@@ -27,7 +27,7 @@ class AttachmentService
     }
 
 	/**
-	 * @param FileBag|UploadedFile[] $files
+	 * @param FileBag|UploadedFile[]|array $files if array it's an assoc array between originalFileName and serverFileName
 	 * @param Arrivage $arrivage
 	 * @param Litige|null $litige
 	 * @param MouvementTraca|null $mvtTraca
@@ -36,18 +36,22 @@ class AttachmentService
         if ($files instanceof FileBag) {
             $files = $files->all();
         }
-		if (!file_exists(self::ATTACHMENT_DIRECTORY)) {
-			mkdir(self::ATTACHMENT_DIRECTORY, 0777);
-		}
-		foreach ($files as $file) {
-			if ($file) {
-				$filename = uniqid() . '.' . $file->getClientOriginalExtension() ?? '';
-				$file->move(self::ATTACHMENT_DIRECTORY, $filename);
 
-				$pj = new PieceJointe();
-				$pj
-					->setOriginalName($file->getClientOriginalName())
-					->setFileName($filename);
+        $isFileName = count($files) > 0 && is_string($files[array_key_first($files)]);
+
+        foreach ($files as $fileIndex => $file) {
+			if ($file) {
+			    $filename = $isFileName
+                    ? $file
+                    : $this->saveFile($file);
+
+                $pj = new PieceJointe();
+                $pj
+                    ->setOriginalName($fileIndex)
+                    ->setFileName($filename);
+
+                $this->em->persist($pj);
+
 				if ($arrivage) {
 					$arrivage->addAttachement($pj);
 				} elseif ($litige) {
@@ -55,10 +59,23 @@ class AttachmentService
 				} elseif ($mvtTraca) {
 					$mvtTraca->addAttachement($pj);
 				}
-				$this->em->persist($pj);
 			}
 		}
 	}
+
+    /**
+     * @param UploadedFile $file
+     * @return array [originalName (string) => filename (string)]
+     */
+	public function saveFile(UploadedFile $file): array {
+        if (!file_exists(self::ATTACHMENT_DIRECTORY)) {
+            mkdir(self::ATTACHMENT_DIRECTORY, 0777);
+        }
+
+        $filename = uniqid() . '.' . $file->getClientOriginalExtension() ?? '';
+        $file->move(self::ATTACHMENT_DIRECTORY, $filename);
+        return [$file->getClientOriginalName() => $filename];
+    }
 
 	/**
 	 * @param PieceJointe $attachment
@@ -75,9 +92,16 @@ class AttachmentService
 		} elseif ($mvtTraca) {
 			$mvtTraca->removeAttachement($attachment);
 		}
-		$path = "../public/uploads/attachements/" . $attachment->getFileName();
-		unlink($path);
-		$this->em->flush();
+
+        $pieceJointeRepository = $this->em->getRepository(PieceJointe::class);
+        $pieceJointeAlreadyInDB = $pieceJointeRepository->findOneByFileName($attachment->getFileName());
+        if (count($pieceJointeAlreadyInDB) === 1) {
+            $path = "../public/uploads/attachements/" . $attachment->getFileName();
+            unlink($path);
+        }
+
+        $this->em->remove($attachment);
+        $this->em->flush();
 	}
 
 }
